[Chore] Extract the FFmpeg ownership wrappers and custom AVIO
Two more responsibilities leave `media/ffi.rs`, one at a time as the
riskiest file in the project deserves.

`ffi::raii` holds the seven wrappers that exist so an early return frees
what was allocated -- `InputContext`, `OutputContext`, `CodecContext`,
`SwrContext`, `ChannelLayout`, `Packet`, `Frame`. They own a raw pointer
and free it in `Drop`; no transcoding logic moved with them.

`ffi::avio` holds the memory-backed `AVIOContext`, so a bundle decoded in
RAM can be transcoded without being written out first. Its read and seek
callbacks are handed to C, which is why `CustomAvio` carries the
lifetime of the slice it lends.

`AudioFifo` did not move, and the attempt is what showed why: it calls
`drain_encoder` as it drains, so it is not an owning wrapper but a piece
of the encode loop that happens to own a FIFO. It stays with the
transcoding logic, with a comment saying so.

Direction is `error` (leaf) <- `avio` <- `raii`, and the file drops from
1 686 lines to 1 315. The transcode drivers, codec configuration and
frame plumbing are still there; they are algorithm, not ownership, and
the tests that would make moving them safe need real media.
